# assets/swagger.json
{
    "info": {
      "title": "API REST documentation",
      "version": "1.0.0",
      "license": {
        "name": "ISC"
      },
      "description": "API REST pour gérer des utilisateurs"
    },
    "basePath": "/api/v1",
    "host": "localhost:8080",
    "tags": [
      {
        "name": "members",
        "description": "Méthodes permettant de gérer les utilisateurs"
      }
    ],
    "paths": {
      "members/{id}": {
        "get": {
          "tags": ["members"],
          "summary": "members/{id}",
          "consumes": [
            "application/json"
          ],
          "parameters": [
            {
              "name": "id",
              "in": "path",
              "required": true
            }
          ]
        },
        "put": {
          "tags": ["members"],
          "summary": "members/{id}",
          "consumes": [
            "application/json"
          ],
          "parameters": [
            {
              "name": "id",
              "in": "path",
              "required": true
            }
          ]
        },
        "delete": {
          "tags": ["members"],
          "summary": "members/{id}",
          "consumes": [
            "application/json"
          ],
          "parameters": [
            {
              "name": "id",
              "in": "path",
              "required": true
            }
          ]
        }
      },
      "members": {
        "get": {
          "tags": ["members"],
          "summary": "members",
          "consumes": [
            "application/json"
          ],
          "parameters": [
            {
              "name": "max",
              "in": "query",
              "required": false
            }
          ]
        },
        "post": {
          "tags": ["members"],
          "summary": "members",
          "consumes": [
            "application/json"
          ],
          "parameters": [
            {
              "name": "name",
              "in": "body",
              "required": true
            }
          ]
        }
      }
    },
    "swagger": "2.0"
  }